requirements.txt
astor==0.8.1
backports.tarfile==1.2.0
bcrypt==4.3.0
build==1.2.2.post1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
cryptography==45.0.3
docutils==0.21.2
gitdb==4.0.12
GitPython==3.1.44
id==1.5.0
idna==3.10
importlib-metadata==8.7.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
jeepney==0.9.0
keyring==25.6.0
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.7.0
nh3==0.2.21
# Editable install with no version control (odooflow==0.1.0)
-e /home/mohammad/Desktop/odoo_projects/odooflow
packaging==25.0
paramiko==3.5.1
pycparser==2.22
pygments==2.19.1
PyNaCl==1.5.0
pyproject-hooks==1.2.0
readme-renderer==44.0
requests==2.32.3
requests-toolbelt==1.0.0
rfc3986==2.0.0
rich==14.0.0
SecretStorage==3.3.3
shellingham==1.5.4
smmap==5.0.2
tomli==2.2.1
tqdm==4.67.1
twine==6.1.0
typer==0.16.0
typing-extensions==4.14.0
urllib3==2.4.0
zipp==3.22.0
